added image for appointment view, appointment view for doctor, added doctor login routes

+++ client/docappointment-ui/src/components/AppointmentBox.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import moment from 'moment';
 import { Button, Popconfirm, message } from 'antd';
+import userIcon from '../images/user.svg';
 
 function AppointmentBox(props) {
 
@@ -15,22 +16,31 @@ function AppointmentBox(props) {
     
     const opposite = loggedInUserRole === 'Patient' ? appointment.doctorUserId : appointment.patientId;
     
-    const button = props.isUpcoming
-        ? <Popconfirm
-            title="Are you sure you want to cancel the appointment?"
-            okText="Yes"
-            cancelText="No"
-            onConfirm={() => props.deleteAppointment(appointment._id)}
-        >
-            <Button type="primary" danger>Cancel</Button>
-        </Popconfirm>
-        : <Button onClick={() => props.setDoctorToBook(appointment.doctorUserId)} type="primary">Book Again</Button>
+    debugger
+    let button = '';
+    if(loggedInUserRole === 'Patient'){
+        button = props.isUpcoming
+            ? <Popconfirm
+                title="Are you sure you want to cancel the appointment?"
+                okText="Yes"
+                cancelText="No"
+                onConfirm={() => props.deleteAppointment(appointment._id)}
+            >
+                <Button type="primary" danger>Cancel</Button>
+            </Popconfirm>
+            : <Button onClick={() => props.setDoctorToBook(appointment.doctorUserId)} type="primary">Book Again</Button>
+
+    }
+        const photo = opposite.photo ? `http://localhost:3038/${opposite.email}/${opposite.photo}` : userIcon;
     return (
         <div
             title={props.isCancelled ? 'This appointment was cancelled' : ''}
             className={props.isCancelled ? 'appointment-box cancelled-appointment' : 'appointment-box'}>
-            <div>
-                <h3>
+            <div className="appointment-box-header">
+                <div>
+                    <img className="header-image" src={photo} alt="user" />
+                </div>
+                <h3 style={{marginLeft: 10}}>
                     {loggedInUserRole === 'Patient' ? `Dr. ${opposite.name}` : opposite.name}
                 </h3>
             </div>
@@ -41,9 +51,15 @@ function AppointmentBox(props) {
                     {appointmentDate.format('hh:mm A')}
                 </div>
             </div>
+            <div style={{display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginTop: 12}}>
+            <div>
+            <i class="fas fa-mobile-alt"></i> <span style={{fontWeight: 500, marginLeft: 4}}>{opposite.phone}</span>
+            </div>
             <div style={{ textAlign: 'right', margin: 10 }}>
 
                 {button}
+
+            </div>
 
             </div>
         </div>
